[Arch] Move constants to shared/lib and document mdx-components
Move src/constants.ts to src/shared/lib/constants.ts and update
all 7 import locations. Add comment to mdx-components.tsx
explaining it must stay at src root per Next.js/contentlayer.
