+++ src/features/home/sections/connect-section.tsx
@@ -1,6 +1,6 @@
 // src/components/sections/home/connect-section.tsx
 import { Card } from "@/shared/components/card";
-import { CONNECT_OFFERINGS, NAVIGATION_ITEMS } from "@/constants";
+import { CONNECT_OFFERINGS, NAVIGATION_ITEMS } from "@/shared/lib/constants";
 import { ArrowUpRight, Check, MessageCircle, User, Users } from "lucide-react";
 
 function getPlatformIcon(platform: string) {
