+++ src/shared/layout/footer.tsx
@@ -1,5 +1,5 @@
 import { Button } from "@/shared/components/button";
-import { SOCIAL_LINKS } from "@/constants";
+import { SOCIAL_LINKS } from "@/shared/lib/constants";
 import { renderSocialIcon } from "@/shared/lib/social-icons";
 import Link from "next/link";
