+++ src/shared/lib/constants.ts
